requirements.txt
Django==1.6.1
South==0.8.4
argparse==1.2.1
django-dajax==0.9.2
django-dajaxice==0.5.5
wsgiref==0.1.2
